Scala 2.13 fixes
Switched from `JavaConverters` to 2.13's `CollectionConverters`.

Widening conversion from `Long` to `Double` is deprecated, so used `toDouble` instead.

Previously `Buffer` could be assigned to `Seq`, because `Seq` was mutable. Now `Seq` is immutable by default, so `toSeq` is needed to create a `Seq` from a `Buffer`.

`Either` is now right-biased, so `.right` is no longer needed.

`.mapValues` is deprecated, need `.view.mapValues` instead. `.toMap` is also needed to get back to a `Map` from a `MapView`.

Auto-application is deprecated, so `()` must be added in some cases.

Deleted `Enumerators` because it wasn't used, and it was using features unsupported in Scala 2.13.